Correções referente ao Módulo de Transporte Escolar - Validando exclusão #458

+++ ieducar/modules/Api/Views/VeiculoController.php
@@ -32,6 +32,7 @@
  * @version   $Id$
  */
 
+require_once 'include/modules/clsModulesItinerarioTransporteEscolar.inc.php';
 require_once 'include/modules/clsModulesVeiculo.inc.php';
 
 require_once 'Portabilis/Controller/ApiCoreController.php';
@@ -124,6 +125,19 @@ class VeiculoController extends ApiCoreController
 
   }      
 
+  protected function validateIfVeiculoIsNotInUse(){
+
+      $it = new clsModulesItinerarioTransporteEscolar();
+      $lista = $it->lista(null,null,null,$this->getRequest()->id);
+      if(is_array($lista) && count($lista)>0){
+        $this->messenger->append('Não é possível excluir um veículo que está vinculada a um itinerário.',
+                                 'error', false, 'error');
+        return false;
+      }else{
+        return true;
+      }
+  }  
+
   protected function get() {
     
       $id                                      = $this->getRequest()->id;
@@ -154,9 +168,15 @@ class VeiculoController extends ApiCoreController
 
       $veiculo = Portabilis_Array_Utils::filter($veiculo, $attrs);
 
-      $veiculo['empresaNome']                   = $this->loadNomeEmpresa($veiculo['empresa']);
+      $veiculo['empresaNome']                   = Portabilis_String_Utils::toUtf8($this->loadNomeEmpresa($veiculo['empresa']));
 
-      $veiculo['motoristaNome']                 = $this->loadNomeMotorista($veiculo['motorista']);
+      $veiculo['motoristaNome']                 = Portabilis_String_Utils::toUtf8($this->loadNomeMotorista($veiculo['motorista']));
+      $veiculo['descricao']                     = Portabilis_String_Utils::toUtf8($veiculo['descricao']);
+      $veiculo['marca']                         = Portabilis_String_Utils::toUtf8($veiculo['marca']);
+      $veiculo['placa']                         = Portabilis_String_Utils::toUtf8($veiculo['placa']);
+      $veiculo['chassi']                        = Portabilis_String_Utils::toUtf8($veiculo['chassi']);
+      $veiculo['descricao_inativo']             = Portabilis_String_Utils::toUtf8($veiculo['descricao_inativo']);
+      $veiculo['observacao']                    = Portabilis_String_Utils::toUtf8($veiculo['observacao']);
     
       return $veiculo;
 
@@ -223,12 +243,19 @@ class VeiculoController extends ApiCoreController
     elseif ($this->isRequestFor('post', 'veiculo'))
       $this->appendResponse($this->post());
 
+    elseif ($this->isRequestFor('delete', 'veiculo')){
+       if($this->validateIfVeiculoIsNotInUse()){
+        $this->appendResponse($this->delete());
+        echo "<script language= \"JavaScript\">
+                location.href=\"intranet/transporte_veiculo_lst.php\";
+              </script>";
+
+        die();
+      }
     // update
-    elseif ($this->isRequestFor('put', 'veiculo'))
+    }elseif ($this->isRequestFor('put', 'veiculo'))
       $this->appendResponse($this->put());
 
-    elseif ($this->isRequestFor('delete', 'veiculo'))
-      $this->appendResponse($this->delete());
     else
       $this->notImplementedOperationError();
   }
